move secret funtions to their own file
